+++ lib/Crypto/SelfTest/PublicKey/test_import_RSA.py
@@ -426,6 +426,11 @@ Lr7UkvEtFrRhDDKMtuIIq19FrL4pUIMymPMSLBn3hJLe30Dw48GQM4UCAwEAAQ==
         self.assertEqual(key.n, self.n)
         self.assertEqual(key.e, self.e)
 
+    def test_import_key(self):
+        """Verify that import_key can be used on bytearrays and memoryviews"""
+        key = RSA.import_key(bytearray(self.rsaPublicKeyDER))
+        key = RSA.import_key(memoryview(self.rsaPublicKeyDER))
+
     def test_exportKey(self):
         key = RSA.construct([self.n, self.e, self.d, self.p, self.q, self.pInv])
         self.assertEqual(key.export_key(), key.exportKey())